Delete old prompt file update code
